# swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signupLocal",
        "summary": "Create patient user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              },
              "examples": {
                "Required_Fields": {
                  "value": {
                    "firstName": "string",
                    "lastName": "string",
                    "email": "string@gmail.com",
                    "password": "string",
                    "gender": "MALE",
                    "birthday": "2000-01-01"
                  },
                  "summary": "Required field"
                },
                "ExpoToken": {
                  "value": {
                    "firstName": "string",
                    "lastName": "string",
                    "email": "string@gmail.com",
                    "password": "string",
                    "birthday": "2000-01-01",
                    "gender": "MALE",
                    "expoToken": "string"
                  },
                  "summary": "Required field + Expo Token"
                },
                "Image": {
                  "value": {
                    "firstName": "string",
                    "lastName": "string",
                    "email": "string@gmail.com",
                    "gender": "MALE",
                    "birthday": "2000-01-01",
                    "password": "string",
                    "image": "string"
                  },
                  "summary": "Required field + Image Base64"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Created account"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "string",
                    "error": "Bad Request"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signin": {
      "post": {
        "operationId": "AuthController_signinLocal",
        "summary": "Sign in and get access and refresh tokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              },
              "examples": {
                "patient": {
                  "value": {
                    "email": "new@gmail.com",
                    "password": "string"
                  },
                  "summary": "Patient login"
                },
                "doctor": {
                  "value": {
                    "email": "doctor@gmail.com",
                    "password": "string"
                  },
                  "summary": "Doctor login"
                },
                "expotoken": {
                  "value": {
                    "email": "doctor@gmail.com",
                    "password": "string",
                    "expoToken": "string"
                  },
                  "summary": "Login expo Token example"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "Incorrect_Password": {
                    "value": {
                      "statusCode": 400,
                      "message": "Incorrect password",
                      "error": "Bad Request"
                    }
                  },
                  "No_User_with_email": {
                    "value": {
                      "statusCode": 400,
                      "message": "No user with current email adress",
                      "error": "Bad Request"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "summary": "Use Refresh token",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized error",
            "content": {
              "application/json": {
                "examples": {
                  "No_Refresh_Token": {
                    "value": {
                      "statusCode": 401,
                      "message": "No Refresh token",
                      "error": "Unauthorized"
                    }
                  },
                  "Incorrect_Refresh_Token_Format": {
                    "value": {
                      "statusCode": 401,
                      "message": "Wrong Refresh Token format",
                      "error": "Unauthorized"
                    }
                  },
                  "No_User": {
                    "value": {
                      "statusCode": 401,
                      "message": "No user with this token",
                      "error": "Unauthorized"
                    }
                  },
                  "No_Match": {
                    "value": {
                      "statusCode": 401,
                      "message": "Incorrect Refresh token",
                      "error": "Unauthorized"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_retrieveUserInfo",
        "summary": "Use Access token get user info",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "email": "string@email.com",
                    "firstName": "string",
                    "lastName": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/articles/get_articles": {
      "get": {
        "operationId": "ArticlesController_get_atricles",
        "summary": "gets all the article in a paginated responses",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "docs": [
                      {
                        "_id": "64478b044db26ae46c8d2389",
                        "title": "string",
                        "date": "2023-04-25T08:10:44.480Z",
                        "content": "string",
                        "__v": 0
                      }
                    ],
                    "totalDocs": 1,
                    "offset": 0,
                    "limit": 10,
                    "totalPages": 1,
                    "page": 1,
                    "pagingCounter": 1,
                    "hasPrevPage": false,
                    "hasNextPage": false,
                    "prevPage": null,
                    "nextPage": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/articles/article/{article_id}": {
      "get": {
        "operationId": "ArticlesController_get_article",
        "summary": "Gets a specific article using id",
        "parameters": [
          {
            "name": "article_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/articles/post": {
      "patch": {
        "operationId": "ArticlesController_post_article",
        "summary": "Post an article",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArticleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull Response",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "msg": "Created Article"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Access",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Forbidden resource",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Articles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Theraline endpoints",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "expoToken": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "gender",
          "birthday"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "expoToken": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
...
      "ArticleDto": {
        "type": "object",
        "properties": {}
      },
      "CreateArticleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ]
      }
    }
  }
}